finishing the app

+++ resolvers/Mutation/usuario.js
@@ -14,7 +14,9 @@ const mutations = {
             }
         })
     },
-    async novoUsuario(_, { dados }) {
+    async novoUsuario(_, { dados }, ctx) {
+        ctx && ctx.validarAdmin()
+
         try {
             const idsPerfis = []
 
@@ -49,7 +51,9 @@ const mutations = {
             throw new Error(e.sqlMessage)
         }
     },
-    async excluirUsuario(_, args) {
+    async excluirUsuario(_, args, ctx) {
+        ctx && ctx.validarAdmin()
+
         try {
             const usuario = await obterUsuario(_, args)
             if (usuario) {
@@ -66,11 +70,14 @@ const mutations = {
 
     },
     async alterarUsuario(_, { filtro, dados }) {
+        ctx && ctx.validarUsuarioFiltro(filtro)
+
         try {
             const usuario = await obterUsuario(_, { filtro })
             if (usuario) {
                 const { id } = usuario
-                if (dados.perfis) {
+                // somente usuários admins podem alterar perfil de outros usuários
+                if (ctx.admin && dados.perfis) {
                     await db('usuarios_perfis')
                         .where({ usuario_id: id }).delete()
